tbi(fiori-gen): consolidate and move getbackendsytemtype fn
