+++ packages/odata-service-inquirer/src/prompts/datasources/sap-system/prompt-helpers.ts
@@ -1,8 +1,5 @@
 import { t } from '../../../i18n';
 import { validateSystemName } from './validators';
-import type { BackendSystem } from '@sap-ux/store';
-
-export type SystemType = 'OnPrem' | 'S4HC' | 'BTP' | undefined;
 
 /**
  * Provides a system name suggestion based on the passed system name and client, validating the name is unique against the secure store.
@@ -35,22 +32,4 @@ async function appendSuffix(systemName: string) {
         suffixNumber++;
     }
     return suffixedSystemName;
-}
-
-/**
- * Determines the backend system type based on the authentication type and service keys (defaults to OnPrem).
- *
- * @param system - the backend system to determine the type for.
- * @returns - the system type
- */
-export function getBackendSystemType(system: BackendSystem): SystemType {
-    let backendSystemType: SystemType;
-    if (system.authenticationType === 'reentranceTicket') {
-        backendSystemType = 'S4HC';
-    } else if (system.serviceKeys) {
-        backendSystemType = 'BTP';
-    } else if (system.authenticationType === 'basic' || system.username) {
-        backendSystemType = 'OnPrem';
-    }
-    return backendSystemType;
 }
