+++ packages/odata-service-inquirer/test/unit/prompts/sap-system/prompt-helpers.test.ts
@@ -1,5 +1,5 @@
 import { initI18nOdataServiceInquirer } from '../../../../src/i18n';
-import { suggestSystemName, getBackendSystemType } from '../../../../src/prompts/datasources/sap-system/prompt-helpers';
+import { suggestSystemName } from '../../../../src/prompts/datasources/sap-system/prompt-helpers';
 import type { BackendSystem } from '@sap-ux/store';
 
 jest.mock('@sap-ux/store', () => ({
@@ -24,22 +24,5 @@ describe('Test prompt-helpers', () => {
         expect(await suggestSystemName('system1', client)).toBe('system1, client 010');
         expect(await suggestSystemName('system1')).toBe('system1 (1)');
         expect(await suggestSystemName('system2')).toBe('system2 (2)');
-    });
-
-    test('Test that the backend system type is determined correctly', () => {
-        const systemWithReentranceTicket = {
-            authenticationType: 'reentranceTicket'
-        };
-        const systemWithServiceKeys = {
-            serviceKeys: '<serviceKeys>'
-        };
-        const systemOnPrem = {
-            username: 'user1',
-            password: 'password1'
-        };
-
-        expect(getBackendSystemType(systemWithReentranceTicket as BackendSystem)).toBe('S4HC');
-        expect(getBackendSystemType(systemWithServiceKeys as BackendSystem)).toBe('BTP');
-        expect(getBackendSystemType(systemOnPrem as BackendSystem)).toBe('OnPrem');
     });
 });
